requirements.txt
streamlit
pandas
jinja2
resend
httpx
python-dotenv
openpyxl
fastapi
uvicorn
pydantic
python-multipart
